Seperate Parser.h and Parser.cpp from query/Plan.cpp
