+++ internal/core/src/query/CMakeLists.txt
@@ -21,6 +21,7 @@ set(MILVUS_QUERY_SRCS
         visitors/VerifyExprVisitor.cpp
         visitors/ExtractInfoPlanNodeVisitor.cpp
         visitors/ExtractInfoExprVisitor.cpp
+        Parser.cpp
         Plan.cpp
         SearchOnGrowing.cpp
         SearchOnSealed.cpp
