+++ internal/core/src/query/Plan.h
@@ -10,9 +10,12 @@
 // or implied. See the License for the specific language governing permissions and limitations under the License
 
 #pragma once
+
 #include <memory>
 #include <string_view>
 #include <string>
+
+#include "PlanImpl.h"
 #include "common/Schema.h"
 #include "pb/plan.pb.h"
 #include "pb/segcore.pb.h"
@@ -50,5 +53,3 @@ int64_t
 GetTopK(const Plan*);
 
 }  // namespace milvus::query
-
-#include "PlanImpl.h"
