+++ internal/core/src/query/PlanProto.h
@@ -9,6 +9,8 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License
 
+#pragma once
+
 #include "common/Schema.h"
 #include "query/PlanNode.h"
 #include "pb/plan.pb.h"
